feat: implement shared Dropdown component
