+++ src/features/Product/components/Comment.module.css
@@ -41,3 +41,24 @@
   font-weight: 400;
   line-height: 1.125rem;
 }
+
+.dropdown ul {
+  width: 139px;
+  height: 92px;
+}
+
+.dropdown li {
+  padding: 12px 0 8px;
+}
+
+.dropdown li:last-child {
+  padding: 8px 0 12px;
+}
+
+.dropdown button {
+  font-size: 1rem;
+  font-weight: 400;
+  line-height: 1.625rem;
+  color: var(--gray500);
+  text-align: center;
+}
