--- src/features/Product/components/Dropdown.module.css
@@ -1,38 +1,0 @@
-.dropdownContainer {
-  position: relative;
-}
-
-.dropdown {
-  position: absolute;
-  border: 1px solid var(--gray200);
-  border-radius: 12px;
-  background-color: #ffffff;
-  width: 139px;
-  height: 92px;
-  right: 0;
-  top: 120%;
-}
-
-.kebab {
-  cursor: pointer;
-}
-
-.dropdown li {
-  width: 100%;
-  cursor: pointer;
-  padding: 12px 0 8px;
-}
-
-.dropdown li:last-child {
-  padding: 8px 0 12px;
-}
-
-.dropdown button {
-  width: 100%;
-  font-size: 1rem;
-  font-weight: 400;
-  line-height: 1.625rem;
-  color: var(--gray500);
-  text-align: center;
-  cursor: pointer;
-}
