--- src/features/ProductList/components/Dropdown.module.css
@@ -1,80 +1,0 @@
-.dropdown {
-  order: 1;
-  position: relative;
-  user-select: none;
-}
-
-.dropdownButton {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 42px;
-  height: 42px;
-  border-radius: 12px;
-  border: 1px solid var(--gray200);
-  cursor: pointer;
-}
-
-.dropdownButton > span {
-  display: none;
-}
-
-.dropdownButton > img {
-  width: 24px;
-  height: 24px;
-}
-
-.dropdownMenu {
-  position: absolute;
-  border: 1px solid var(--gray200);
-  border-radius: 12px;
-  background-color: #ffffff;
-  width: 130px;
-  height: 84px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  right: 0;
-  top: 120%;
-}
-
-.dropdownMenu li {
-  width: 100%;
-  text-align: center;
-  cursor: pointer;
-  font-size: 1rem;
-  font-weight: 400;
-  line-height: 1.625rem;
-  color: var(--gray800);
-  padding: 9px 0 7px;
-}
-
-.dropdownMenu button {
-  width: 100%;
-  cursor: pointer;
-}
-
-.dropdownMenu li:first-child {
-  border-bottom: 1px solid var(--gray200);
-}
-
-@media screen and (min-width: 768px) {
-  .dropdown {
-    order: 0;
-  }
-
-  .dropdownButton {
-    justify-content: space-evenly;
-    width: 130px;
-    height: 42px;
-  }
-
-  .dropdownButton > img {
-    content: url(assets/images/ic_arrow_down.svg);
-  }
-
-  .dropdownButton > span {
-    display: inline-block;
-  }
-}
