+++ src/features/ProductList/components/ProductList.module.css
@@ -66,6 +66,43 @@
   row-gap: 32px;
 }
 
+.dropdown {
+  order: 1;
+  user-select: none;
+}
+
+.dropdown > div {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 42px;
+  height: 42px;
+  border-radius: 12px;
+  border: 1px solid var(--gray200);
+}
+
+.dropdown ul {
+  width: 130px;
+  height: 84px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.dropdown li {
+  text-align: center;
+  font-size: 1rem;
+  font-weight: 400;
+  line-height: 1.625rem;
+  color: var(--gray800);
+  padding: 9px 0 7px;
+}
+
+.dropdown li:first-child {
+  border-bottom: 1px solid var(--gray200);
+}
+
 @media screen and (min-width: 768px) {
   .section {
     margin-bottom: 40px;
@@ -91,6 +128,24 @@
     grid-template-columns: repeat(3, 1fr);
     column-gap: 16px;
     row-gap: 40px;
+  }
+
+  .dropdown {
+    order: 0;
+  }
+
+  .dropdown > div {
+    justify-content: space-evenly;
+    width: 130px;
+    height: 42px;
+  }
+
+  .dropdown img {
+    content: url(assets/images/ic_arrow_down.svg);
+  }
+
+  .dropdown span {
+    display: inline-block;
   }
 }
